requirements.txt
python-telegram-bot==22.1
python-dotenv
requests